refactor: move order-related functions to a dedicated orders directory and update imports
