+++ apps/vendor-okx/src/orders/cancelOrder.ts
@@ -3,7 +3,7 @@ import { decodePath } from '@yuants/utils';
 import { ICredential, postTradeCancelOrder } from '../api/private-api';
 
 export const cancelOrder = async (credential: ICredential, order: IOrder) => {
-  const [, instId] = decodePath(order.product_id);
+  const [instId] = decodePath(order.product_id).slice(-1);
   const res = await postTradeCancelOrder(credential, {
     instId,
     ordId: order.order_id!,
